Turn feed into a lazy computed property
This isn't what I wanted to do, but I wasn't able to hit the EntryStore
and find the right feed inside of that init. I was getting an exception
about the struct being captured by a closure before being initialized.
Fair enough.

So, I pulled back on the whole customized decoding and instead went with
a lazy computed property. We'll see how well this works long-term.